==> analysis/02_kidswords_dataprep.R <==
# ---------------------------
# KidsWords project
# Data preparation-02
# Thomas Klee
# Created: 2017-04-17
# Updated: 2019-04-19
# ---------------------------

# This script: 
# defines the word_lookup data frame; 
# cleans the PQ data frame, renames variables and declares some as factors;
# adds a new variable to CDI data frame indicating session order of 
# cross-sectional and longitudinal data; 
# calculates child's age as a period and adds it to CDI data frame;
# calculates child's age in months as an integer and adds it.

# The script requires these data files: 
# (1) data_CDI.csv and (2) data_PQ.csv were generated by 
# 01_kidswords_dataprep.R in order to anonymize the 
# original data files and remove records from the website's 
# data entry test phase and the Learning to Talk project. 
# (3) cdi_lookup.csv labels each item on the CDI and
# describes some linguistic features of the CDI words.
# (4) Parent education level was recoded by hand into summary
# categories used by Statistics NZ (in KidsWords_peduc_20170904.csv).

# The script makes changes to several fields of the parent questionnaire.
# The new information resulted from contacting parents about missing data
# or apparently incorrect data entered by them on the website.

# The script also cleans birth weights recorded in the PQ.
# Note that some decision rules were used to code some data as missing (NA).
# These can be fine-tuned in the future by future researchers.

# The script merges the lookup and CDI data frames and creates 
# a new variable ("resp" in CDI_words data frame) containing 1s and 0s 
# instead of "Says" and NA. The "resp" variable also contains 1s and 0s 
# for CDI Section E items originally coded with values from 0-2.

# The script creates CDI_text, a data frame consisting of CDI items with
# open-ended, text-based responses.

# The script produces long (CDI_words) and wide (CDI, CDI_wide) data frames.

# load packages
library(tidyverse)
library(lubridate)

# load data from the CDI and the parent questionnaire
CDI <- read_csv("data/data_CDI.csv")
PQ <- read_csv("data/data_PQ.csv")

# load lookup table of CDI vocabulary characteristics
word_lookup<- read_csv("data/cdi_lookup.csv")

# load recoded parent education data from the PQ
PQed <- read_csv("data/KidsWords_peduc_20170904.csv")

# replace incorrect birthdates
CDI <- mutate(CDI, BIRTHDAY = replace(BIRTHDAY, CHILD_ID == 22826, "16-03-12"))
CDI <- mutate(CDI, BIRTHDAY = replace(BIRTHDAY, CHILD_ID == 20574, "23-03-10")) # from KidsWords participants database

# locate records with missing sex (i_1_2) data in PQ dataframe
PQ[is.na(PQ$i_1_2), ]

# replace missing sex values in PQ dataframe based on values in CDI dataframe 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 20559, "Boy")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 20955, "Girl")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 21379, "Girl")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 21550, "Boy")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 21856, "Boy")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 22373, "Boy")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 22788, "Boy")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 22944, "Boy")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 23571, "Boy")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 23582, "Girl")) 
PQ <- mutate(PQ, i_1_2 = replace(i_1_2, SCL_ID == 24460, "Boy")) 

# locate records with missing birth order data in PQ
PQ[is.na(PQ$i_1_3), ]

# replace missing birth order values in PQ dataframe based on follow-up with parents
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 20565, "2nd")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 20955, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 20957, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 21786, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 21856, "2nd")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 22779, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 23582, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 23786, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 23884, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 23976, "1st")) 
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 24161, "1st"))
PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 24280, "4th or more")) 
# PQ <- mutate(PQ, i_1_3 = replace(i_1_3, SCL_ID == 22325, "")) 

# locate records with missing region data in PQ
PQ[is.na(PQ$i_1_33), ]

# replace missing region values in PQ based on follow-up with parents
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 20889, "Auckland")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 21302, "Auckland")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 21615, "Marlborough")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 21856, "Canterbury")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 22209, "Wellington")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 22660, "Otago")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 22758, "Wellington")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 23007, "Waikato")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 23395, "Auckland")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 23582, "Marlborough")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 23723, "Auckland")) 
PQ <- mutate(PQ, i_1_33 = replace(i_1_33, SCL_ID == 24292, "Auckland")) 

# locate records with missing twin data in PQ
PQ[is.na(PQ$i_1_7), ] %>% 
  print(n = 30)

# replace missing twin data in PQ based on follow-up with parents
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 20554, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 20583, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 20589, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 20598, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 20942, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 21001, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 21014, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 21222, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 21548, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 21856, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 22199, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 22201, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 22322, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 22628, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 22635, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 22907, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 22912, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23259, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23437, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23560, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23582, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23682, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23687, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23721, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23835, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 23976, "no")) 
PQ <- mutate(PQ, i_1_7 = replace(i_1_7, SCL_ID == 24212, "no")) 

# rename two character string variables 
CDI <- rename(CDI, DOB = BIRTHDAY) 
CDI <- rename(CDI, DOS = FILLEDOUT) # date CDI was completed (session date)

# convert character strings to date class using lubridate
CDI$DOB <- dmy(CDI$DOB)
CDI$DOS <- dmy(CDI$DOS) 

# calculate interval between dates, expressed as a time span
CDI$span <- interval(CDI$DOB, CDI$DOS)

# convert interval to a period, expressed in months and days
CDI$span <- as.period(CDI$span, unit = "months") 

# calculate child's age in months
# extract number of months (as an integer)
CDI$camos <- month(CDI$span)

CDI <- mutate(CDI, cadays = DOS - DOB)

# convert child's age in days to an integer
CDI$cadays <- as.integer(CDI$cadays)

# create "session" variable and assign ranks to its values based session date
CDI <- CDI %>% group_by(CHILD_ID)
CDI <- CDI %>% mutate(session = rank(DOS))

# convert session variable to a factor
CDI$session <- factor(CDI$session)

# sort records by child ID and session number (base R syntax used instead of tidyverse for this)
CDI <- CDI[order(CDI$CHILD_ID, CDI$session),]

# rename variables using dplyr format (new_name = current_name)
CDI <- rename(CDI, PID = CHILD_ID)
PQ <- rename(PQ,
             PID = SCL_ID,
             #            cname = i_1_1, # removed
             csex = i_1_2,
             cbirth_order = i_1_3,
             cbirth_weight_g = i_1_4,
             cbirth_weight_lb = i_1_5,
             cbirth_weight_oz = i_1_6,
             ctwin = i_1_7,
             cpremature = i_1_8,
             cprem_wks = i_1_9,
             chealth_probs = i_1_10,
             chealth_what = i_1_11,
             csiblings = i_1_12,
             cborn_in_nz = i_1_13,
             ccountry_where = i_1_14,
             ctime_in_nz = i_1_15,
             cethnicity_nz = i_1_16,
             cethnicity_other = i_1_17,
             cmain_lang = i_1_18,
             cother_langs = i_1_19,
             cwhich_langs = i_1_20,
             cdaycare = i_1_21,
             cdaycare_hrs = i_1_22,
             PQ3_hear_conc = i_1_23,
             PQ1_lang_conc = i_1_24,
             PQ2_comm_conc = i_1_25,
             why_conc = i_1_26,
             discuss_conc = i_1_27,
             #            home_phone = i_1_28, # removed
             #            cell_phone = i_1_29, # removed
             #            email_address = i_1_30, # removed
             family_hist = i_1_31,
             family_hist_who = i_1_32,
             region = i_1_33,
             #            pname = i_2_1, # removed
             prelation = i_2_2,
             pbirth_place = i_2_3,
             pother_place = i_2_4,
             parrived_nz = i_2_5,
             pethnicity_nz = i_2_6,
             pethnicity_other = i_2_7,
             psec_qual = i_2_8,
             pother_qual = i_2_9,
             pother_qual_what = i_2_10,
             poccupation = i_2_11
)

# declare some variables as nominal or ordered factors
PQ$csex <- factor(PQ$csex)
PQ$cbirth_order <- ordered(PQ$cbirth_order)
PQ$ctwin <- factor(PQ$ctwin)
PQ$cpremature <- factor(PQ$cpremature)
PQ$chealth_probs <- factor(PQ$chealth_probs)
PQ$cborn_in_nz <- factor(PQ$cborn_in_nz)
PQ$ccountry_where <- factor(PQ$ccountry_where)
PQ$cethnicity_nz <- factor(PQ$cethnicity_nz)
PQ$cethnicity_other <- factor(PQ$cethnicity_other)
PQ$cmain_lang <- factor(PQ$cmain_lang)
PQ$cother_langs <- factor(PQ$cother_langs)
PQ$cmain_lang <- factor(PQ$cmain_lang)
PQ$cother_langs <- factor(PQ$cother_langs)
PQ$cwhich_langs <- factor(PQ$cwhich_langs)
PQ$cdaycare <- factor(PQ$cdaycare)
PQ$PQ3_hear_conc <- factor(PQ$PQ3_hear_conc)
PQ$PQ1_lang_conc <- factor(PQ$PQ1_lang_conc)
PQ$PQ2_comm_conc <- factor(PQ$PQ2_comm_conc)
PQ$family_hist <- factor(PQ$family_hist)
PQ$region <- factor(PQ$region)
PQ$prelation <- factor(PQ$prelation)
PQ$pbirth_place <- factor(PQ$pbirth_place)
PQ$pother_place <- factor(PQ$pother_place)
PQ$pethnicity_nz <- factor(PQ$pethnicity_nz)
PQ$pethnicity_other <- factor(PQ$pethnicity_other)
PQ$psec_qual <- factor(PQ$psec_qual)

word_lookup$CDI_part <- factor(word_lookup$CDI_part)
word_lookup$word_class <- factor(word_lookup$word_class)
word_lookup$syl_structure <- factor(word_lookup$syl_structure)

# recode variable values for consistency using dplyr's old = "new" format;
# multiword phrases on left side of "=" are correctly surrounded by quote marks
PQ$prelation <- recode(PQ$prelation, 
                       Mother = "mother", Mum = "mother", mum = "mother", 
                       mummy = "mother", Mummy = "mother",
                       "Mumma!" = "mother", "Mum (one of two)" = "mother",
                       mothet = "mother", Mothet = "mother", Mothers = "mother",
                       "mother to olivia" = "mother",
                       "Im her mother" = "mother", "I am her mother" = "mother",
                       "Im her Mother" = "mother", "I am his mother." = "mother",
                       "I am his mother" = "mother", "His mother." = "mother",
                       "his mother" = "mother", "Her mummy :)" = "mother",
                       "Her mother" = "mother", "her mother" = "mother",
                       "Grandmother and part time carer" = "grandmother",
                       "Grandmother" = "grandmother", "Foster mum" = "foster mother",
                       MOther = "mother", Monther = "mother", MOTHER = "mother",
                       "mother (and stand in for father)" = "mother",
                       mama = "mother", Mama = "mother", mom = "mother",
                       Melissa = "mother", 
                       "Bens Mum" = "mother", Mothe = "mother",
                       Morher = "mother", Moither = "mother",
                       mither = "mother", Māmā = "mother",
                       Father = "father", daddy = "father", 
                       "Mother / Father" = "mother and father",
                       "Mother/Son" = "mother", "Mother Daughter" = "mother",
                       "Mother and son" = "mother", "Mother and daughter" = "mother",
                       "Mother / Son" = "mother", "Mother / Father" = "mother and father",
                       Parents = "mother and father", parents = "mother and father",
                       Parent = "mother or father", "Dad & Mum" = "mother and father",
                       close = "unknown", yes = "unknown", Yes = "unknown",
                       Son = "mother or father", daughter = "mother or father",
                       Daughter = "mother or father", 
                       "He is my son" = "mother or father", monther = "mother"
                       )

PQ$cmain_lang <- recode(PQ$cmain_lang,
                        english = "English",
                        Englinsh  = "English",
                        englinsh  = "English",
                        Emglish = "English",
                        emglish = "English",
                        eglish = "English",
                        Eglish = "English",
                        Eng = "English",
                        Engish = "English",
                        Engliah  = "English",
                        Englisch = "English",
                        englich  = "English",
                        Engligh  = "English",
                        Englih  = "English",
                        "Engllish" = "English",
                        "Enlish" = "English",
                        "Enhlish" = "English",
                        ENGLISH = "English",
                        "english (nz)" = "English",
                        "English / Sign" = "English",
                        "English. baby sign" = "English",
                        "NZEnglish" = "English",
                        "New Zealand English" = "English",
                        "NZSL-supported English (she uses both equally)" = "English",
                        "english / maori" = "English"
# the following records are noted but not changed 
# since these children are apparently being raised in bilingual households
#                       "Englisch/spanish" = "English", 
#                        "combination of English and Bisaya (one of the many dialects in Philippines aside from Tagalog)" = "English"
#                        "English & indonesian" = "English",
#                        "english and german equally" = "English",
#                        "English and Gujarati are being taught simultaneously" = "English",
#                        "English and Indonesian" = "English",
#                        "English and Romanian" = "English",
#                        "english and turkish" = "English",
#                        "English/ French" = "English",
#                        "English/Dutch" = "English",
#                        "English/Romanian (50/50..sort of....)" = "English",
#                        "German and English equal" = "English",
#                        "marathi and english" = "English", 
#                        "Russian with mother. English with father" = "English"
                        )

# replace missing responses (NA) in PQ$cmain_lang with "English" since no reason to assume otherwise
PQ$cmain_lang <- PQ$cmain_lang %>% 
  replace_na("English") 

# create dichotomous variable to indicate whether child's main language environment is monolingual 
PQ$mono_env <- factor(ifelse(PQ$cmain_lang == "English", "yes", "no"))

# select relevant variables from PQed
PQed <- select(PQed, PID, peduc)

# add peduc variable to PQ
PQ <- left_join(PQ, PQed)  # dplyr function

# declare variables as an ordered factor
PQ$peduc <- ordered(PQ$peduc)

# declare baseline (reference) category for csex
# for calculating odds ratios 
PQ$csex <- relevel(PQ$csex, "Girl")

# create single birth weight variable (gram4) from 3 PQ variables

# create new variables from old by removing alphabetic characters
PQ$gram1 <- str_replace_all(PQ$cbirth_weight_g, "[:alpha:]", "")
PQ$lb <- str_replace_all(PQ$cbirth_weight_lb, "[:alpha:]", "")
PQ$oz <- str_replace_all(PQ$cbirth_weight_oz, "[:alpha:]", "")

# change class of character strings 
PQ$gram1 <- as.numeric(PQ$gram1)
PQ$lb <- as.numeric(PQ$lb)
PQ$oz <- as.numeric(PQ$oz)

# convert lb and oz to grams
PQ$lb2 <- PQ$lb * 453.59237
PQ$oz2 <- PQ$oz * 28.34952312

# PQ <- select(PQ, PID, starts_with("cbirth_weight"), gram1, lb, oz, lb2, oz2, starts_with("cprem"))

# where lb was entered but not oz in the PQ, convert NAs to 0
PQ$oz2 <- (ifelse(PQ$lb2 > 0 & is.na(PQ$oz2), 0, PQ$oz2))

# add these together
PQ$gram2 <- PQ$lb2 + PQ$oz2

# if gram1 and gram2 > 0, use gram1
PQ$gram3 <- ifelse((PQ$gram1 > 0 & PQ$gram2 > 0), PQ$gram1, PQ$gram2)

# reorder for easier reading
# PQ <- select(PQ, PID, starts_with("cbirth_weight"), gram1, gram2, gram3, lb, oz, lb2, oz2, starts_with("cprem"))

# combine two variables into a new one
PQ$gram4 <- ifelse((PQ$gram2 > 0 & is.na(PQ$gram1)), PQ$gram2, NA)
PQ$gram4 <- ifelse((PQ$gram1 > 0 & is.na(PQ$gram2)), PQ$gram1, PQ$gram4)

# declare variable to be numeric
# PQ$gram4 <- as.numeric(PQ$gram4)

# reorder for easier reading
# PQ <- select(PQ, PID, starts_with("cbirth_weight"), gram1, gram2, gram3, gram4, lb, oz, lb2, oz2, starts_with("cprem"))

# use metric weight if both metric and imperial were reported
PQ$gram4 <- ifelse(PQ$gram3 > 0 & is.na(PQ$gram4), PQ$gram3, PQ$gram4)

# Some respondents reported birthweight in kg, so
# convert kg to g for any value < 10
PQ$gram4 <- ifelse(PQ$gram4 < 10, PQ$gram4 * 1000, PQ$gram4)
PQ$gram4 <- ifelse(PQ$gram4 >= 10 & PQ$gram4 < 100, PQ$gram4 * 100, PQ$gram4)
# the next command may inadvertently increase bw's in the hundreds
# PQ$gram4 <- ifelse(PQ$gram4 >= 100 & PQ$gram4 < 1000, PQ$gram4 * 10, PQ$gram4)

# hand-calculate remaining birthweights:
# find rows with missing gram data 
missing_bw <- PQ[is.na(PQ$gram4), ] %>% 
  print(n = 80)

# save as CSV file for printing
write.csv(missing_bw, file = "data/missing_bw.csv")

# for each missing birthweight (NA)
# hand-calculate it in grams using conversion formulas above
PQ$gram4 <- ifelse(PQ$PID == 20758 & is.na(PQ$gram4), 3515, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 20761 & is.na(PQ$gram4), 3799, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 20901 & is.na(PQ$gram4), 3430, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 20912 & is.na(PQ$gram4), 3200, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 20990 & is.na(PQ$gram4), 3714, PQ$gram4)
PQ$gram4 <- ifelse(PQ$PID == 21142 & is.na(PQ$gram4), 4337, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21151 & is.na(PQ$gram4), 4196, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21359 & is.na(PQ$gram4), 3374, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21379 & is.na(PQ$gram4), 3714, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21451 & is.na(PQ$gram4), 3685, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21459 & is.na(PQ$gram4), 3515, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21526 & is.na(PQ$gram4), 3941, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21701 & is.na(PQ$gram4), 4026, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21727 & is.na(PQ$gram4), 3005, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 21757 & is.na(PQ$gram4), 3990, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22197 & is.na(PQ$gram4), 3033, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22212 & is.na(PQ$gram4), 4479, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22388 & is.na(PQ$gram4), 3402, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22429 & is.na(PQ$gram4), 3427, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22551 & is.na(PQ$gram4), 3572, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22587 & is.na(PQ$gram4), 3075, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22639 & is.na(PQ$gram4), 3657, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22797 & is.na(PQ$gram4), 3912, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22886 & is.na(PQ$gram4), 3742, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22968 & is.na(PQ$gram4), 4366, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 22980 & is.na(PQ$gram4), 3203, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23077 & is.na(PQ$gram4), 3147, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23111 & is.na(PQ$gram4), 4196, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23120 & is.na(PQ$gram4), 3203, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23121 & is.na(PQ$gram4), 3600, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23225 & is.na(PQ$gram4), 4281, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23249 & is.na(PQ$gram4), 4082, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23390 & is.na(PQ$gram4), 3175, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23402 & is.na(PQ$gram4), 3600, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23437 & is.na(PQ$gram4), 4423, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23475 & is.na(PQ$gram4), 5250, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23532 & is.na(PQ$gram4), 3350, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23569 & is.na(PQ$gram4), 3629, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23573 & is.na(PQ$gram4), 3345, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23616 & is.na(PQ$gram4), 1250, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23627 & is.na(PQ$gram4), 4050, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23651 & is.na(PQ$gram4), 4196, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23687 & is.na(PQ$gram4), 3515, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23693 & is.na(PQ$gram4), 3827, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23740 & is.na(PQ$gram4), 3572, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23770 & is.na(PQ$gram4), 4167, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23814 & is.na(PQ$gram4), 3289, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23882 & is.na(PQ$gram4), 4026, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23892 & is.na(PQ$gram4), 3175, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23919 & is.na(PQ$gram4), 3714, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 23946 & is.na(PQ$gram4), 4026, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 24244 & is.na(PQ$gram4), 3033, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 24249 & is.na(PQ$gram4), 3970, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 24261 & is.na(PQ$gram4), 3898, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 24465 & is.na(PQ$gram4), 3572, PQ$gram4) 
PQ$gram4 <- ifelse(PQ$PID == 24486 & is.na(PQ$gram4), 4026, PQ$gram4) 

# replace mis-calculated birth weights with hand-calculated values
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 20893, 2268))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21405, 3062))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21608, 4479))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21642, 1300))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21751, 1170))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21791, 3658))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22238, NA)) # data entered uninterpretable
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22301, 1446))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22565, 2325)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22734, 1100)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22832, 2000)) # 20000g reported 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22841, 3062)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22877, 3912)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22972, 3714)) # 8lbs, 3oz
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 23178, 4706)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 23277, 1100))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 23304, 4337)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 24047, NA)) # data entered uninterpretable
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 24052, 4196)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 24059, 3912)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 24169, 2920)) 
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 24450, 3402)) 

# identify reported birth weights > 5500 grams and convert
# them to NA on the assumption that they were incorrectly entered
# on the PQ
# high_bw <- filter(PQ, gram4 > 5500)
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 20791, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21581, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21603, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21661, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21711, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21787, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 23219, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 23357, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 23714, NA))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 24391, NA))

# identify birth weights out of those remaining where both metric
# and imperial were reported on the PQ and hand-calculate gram4
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 21481, 3770))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22199, 2523))
PQ <- mutate(PQ, gram4 = replace(gram4, PID == 22760, 4281))

# identify those with reported birth weights < 400 grams 
# and no reported prematurity and convert gram4 to NA 
# on the assumption that they were incorrectly entered on the PQ
PQ <- mutate(PQ, gram4 = replace(gram4, gram4 < 400 & cpremature == "no", NA))

# identify remaining records with unlikely reported birth weights
# and convert them to NA
# low_bw <- filter(PQ, gram4 < 400 & !is.na(cprem_wks)) # test works
PQ <- mutate(PQ, gram4 = replace(gram4, gram4 < 400 & !is.na(cprem_wks), NA))

# round to nearest gram (do this as lasts step, so move this command)
PQ$gram4 <- round(PQ$gram4, digits = 0)

# rename gram4 for final data frame
PQ <- rename(PQ, cbirth_weight_grams = gram4)

# remove temporary object
rm(missing_bw)

# create 2nd data frame from CDI before converting CDI to long-format
CDI_wide <- CDI

# convert CDI data frame from wide- to long-format
CDI_long <- 
  CDI %>%
  gather(key = itemID, value = response, i_1_1:i_32_1, factor_key = TRUE)

# select variables from CDI data frames that don't also occur in PQ
CDI_long <- select(CDI_long, PID, DOB, DOS, session, camos, cadays, 
              itemID, response)
CDI_wide <- select(CDI_wide, PID, DOB, DOS, session, camos, cadays) 

# merge CDI_long and word_lookup data frames
CDI_words <- merge(CDI_long, word_lookup, by = "itemID")

# select csex from PQ and add to CDIwords
PQ_csex <- select(PQ, PID, csex)
CDI_words <- merge(CDI_words, PQ_csex, by = "PID")
rm(PQ_csex)

# create new (duplicate) variable
CDI_words <- mutate(CDI_words, resp = response)

# filter out open-ended response items (30 & 32) and put in new data frame
CDI_text <- CDI_words %>% 
  filter(field == 30 | field == 32) %>% 
  select(PID, session, field, CDI_item, itemID, response)

# delete these text records from CDI_words
CDI_words <- CDI_words %>% 
  filter(field != 30 & field != 32) 

# recode text responses into numeric code in 3 CDI items 
CDI_words$resp <- ifelse(CDI_words$resp == "Often", "2", CDI_words$resp)
CDI_words$resp <- ifelse(CDI_words$resp == "Sometimes", "1", CDI_words$resp)
CDI_words$resp <- ifelse(CDI_words$resp == "Not Yet", "0", CDI_words$resp)

# recode values of new variable (resp) in order to count
# "1"s and "0"s instead of number of rows = "Says".
# Any child having no words checked on CDI will have 
# vocabulary size correctly calculated as 0.
CDI_words$resp <- ifelse(CDI_words$resp == "Says", "1", CDI_words$resp)
CDI_words$resp <- ifelse(is.na(CDI_words$resp), 0, CDI_words$resp) 

# CDI Section E: Complexity
# find records in Section E (field = 31) with resp = 1, and replace with 0;
# find records in Section E with resp = 2 and replace with 1;
# recoding of original "response" variable containing NAs (to resp = 0) was done above
# from https://stackoverflow.com/questions/40705807/replace-values-on-condition-in-r
CDI_words[CDI_words$field == 31 & CDI_words$resp == "1", c("resp")] <- "0"
CDI_words[CDI_words$field == 31 & CDI_words$resp == "2", c("resp")] <- "1"

# declare "resp" variable to be an integer
CDI_words$resp <- as.integer(CDI_words$resp)

sessionInfo()
