==> analysis/04_kidswords_stats.R <==
# ---------------------------
# KidsWords Project
# Summary statistics
# Thomas Klee
# Created: 2017-06-22
# Updated: 2017-09-19
# ---------------------------

# This script constructs summary tables and graphs.
# R code is supplemented by LaTeX code so that analyses 
# can be used in manuscripts etc.

library(tidyverse)
library(xtable)
library(RColorBrewer)
library(Hmisc)
library(vioplot)
library(ggplot2)
library(ggthemes)
library(ggridges)

# load CDI and parent questionnaire data
# generated by 03_kidswords_dataprep.csv

# Note that read.csv function is used instead of read_csv
# to maintain variable types (eg factor, character, integer)
# defined in 02_kidswords_dataprep.R

CDIPQ <- read.csv("data/data_CDIPQ.csv") 

# summarize demographic characteristics ---------

# for the full sample,
# tabulate sample size of each age x sex cell and add margin totals
tbl_agesex1 <- addmargins(table(CDIPQ$camos, CDIPQ$csex, dnn = c("Age", "")))
colnames(tbl_agesex1) <- c("Boys", "Girls", "Total")
tbl_agesex1

# the "Age" column header gets lost when adding latex commands; otherwise ok
print(xtable(tbl_agesex1, digits = 0, align = "lrrr", caption = 'Sample size by age and sex', 
             label = "tbl:agesex1"), caption.placement = 'top', booktabs = TRUE)

# -----------------------------------------------
# select data for the cross-sectional (normative) sample
CDIPQxs <- CDIPQ %>% 
  filter(session == 1,  camos >= 16 & camos <= 30)

# tabulate size of each age x sex cell and add margin totals
tbl_agesex2 <- addmargins(table(CDIPQxs$camos, CDIPQxs$csex, dnn = c("Age", "")))
colnames(tbl_agesex2) <- c("Boys", "Girls", "n")
tbl_agesex2

# "Age" column header added above gets lost when adding latex commands; otherwise ok
print(xtable(tbl_agesex2, digits = 0, align = "lrrr", caption = 'Sample size by age and sex', 
             label = "tbl:agesex2"), caption.placement = 'top', booktabs = TRUE)

# -----------------------------------------------
# tabulate sample size by birth order and sex
bosex_counts <- table(CDIPQxs$cbirth_order, CDIPQxs$csex)
bosex_rowcounts <- rowSums(bosex_counts) # calculate row totals

# convert counts to percentages
bosex_props <- (prop.table(table(CDIPQxs$cbirth_order)) * 100)

# combine population and sample data, name columns, create latex table 
tbl_bosex <- cbind(bosex_counts, bosex_rowcounts, bosex_props)
colnames(tbl_bosex) <- c("Boys", "Girls", "Total", "%")
tbl_bosex

print(xtable(tbl_bosex,  
             digits = c(0, 0, 0, 0, 1), align = "lrrrr", 
             label = "tbl:bosex"), booktabs = TRUE)

# -----------------------------------------------
# compare regional distribution of sample and NZ population
statsnz <- read_csv("data/NZBirths_20140914_NA_omit.csv")

NZbirths <- table(statsnz$region, statsnz$runAve)
NZbirths <- round((statsnz$runAve * 100), digits = 1)

# Calculate number of births per region in the KidsWords sample
region_totals <- table(CDIPQxs$region)

# Convert frequencies to percentages
region_props <- (prop.table(table(CDIPQxs$region)) * 100)

# Combine population and sample data, name columns, create latex table 
region_tbl <- with(CDIPQxs, cbind(region_totals, region_props, NZbirths))
colnames(region_tbl) <- c("n", "Sample (%)", "Population (%)")
region_tbl

print(xtable(region_tbl, digits = c(0, 0, 1, 1), align = "lrrr", 
             caption = 'Sample size by region in relation to distribution of 
             births in the population; ignore first row for now',
             label = "tbl:regions"), caption.placement = 'top', booktabs = TRUE)

# -----------------------------------------------
# aggregate peduc codes (0-9) into a new variable (pedcat)
# using Statistics NZ categories and declare as an ordered factor:
# 0 = no qualification
# 1 = level 1-2 certificates (some secondary qualification)
# 2 = level 3-4 certificates and level 5-6 diplomas (secondary education certs and dips)
# 3 = level 7-10 undergrad and postgrad degrees and grad and postgrad 
# certs and dips (university degree)
# 11 = level 11 overseas secondary school qualification
CDIPQxs$pedcat <- recode(CDIPQxs$peduc,
                         '0' = "0",
                         '1' = "1",
                         '2' = "1",
                         '3' = "2",
                         '4' = "2",
                         '5' = "2",
                         '6' = "2",
                         '7' = "3",
                         '8' = "3",
                         '9' = "3",
                         '10' = "3",
                         '11' = "2"
)

CDIPQxs$pedcat <- ordered(CDIPQxs$pedcat)

CDIPQxs$camos <- ordered(CDIPQxs$camos)

# plot word totals by maternal education --------

# density ridgeline plot
ggplot(CDIPQxs, aes(x = wordtotal, y = pedcat)) + 
  geom_density_ridges(scale = 0.85) +
  scale_x_continuous(limits = c(0, 700)) +
  theme_few() +
  scale_colour_few(palette = "dark")  

# notched box plot using ggplot
ggplot(CDIPQxs, aes(pedcat, wordtotal)) + 
  stat_boxplot(notch = TRUE) + 
  theme_few()

# notched box plot using base R
boxplot(wordtotal ~ pedcat, data = CDIPQxs, 
        notch = TRUE, col = "linen", range=1.5,
        names = c("A", "B", "C", "D"),
#        names = c("some secondary qualification", 
#                  "secondary education certs and dips", 
#                  "undergrad and postgrad degrees and grad and postgrad",
#                  "university degree"),
        xlab = "Maternal Education Attained", 
        ylab = "Vocabulary Size in Words")

# -----------------------------------------------
# calculate percentage of each education level in the sample
cat0 <- filter(CDIPQxs, pedcat == 0)
cat0pct <- nrow(cat0) / nrow(CDIPQxs)

cat1 <- filter(CDIPQxs, pedcat == 1)
cat1pct <- nrow(cat1) / nrow(CDIPQxs)

cat2 <- filter(CDIPQxs, pedcat == 2)
cat2pct <- nrow(cat2) / nrow(CDIPQxs)

cat3 <- filter(CDIPQxs, pedcat == 3)
cat3pct <- nrow(cat3) / nrow(CDIPQxs)

# for the purpose of comparison, plot summary data for CDI word totals
# and grammatical complexity in different ways: 
# notched boxplots, box-percentile plots, violin plots, density ridgeline plots

# -----------------------------------------------
# create notched boxplots of CDI word total for each month
boxplot(wordtotal ~ camos, data = CDIPQxs, 
        notch=TRUE, col = "linen", range=1.5,
        # main = "Vocabulary by Age", 
        xlab = "Age in Months", 
        ylab = "Vocabulary Size in Words")

# calculate mean number of words for age group
mean_words <- aggregate(wordtotal ~ camos, CDIPQxs, mean)

# add mean number of words to graph
points(mean_words$wordtotal, pch = "*")

# -----------------------------------------------
# create box-percentile plots of CDI word total for each month
bpplot(split(CDIPQxs$wordtotal, CDIPQxs$camos),
       xlab = "Age in Months", 
       ylab = "Vocabulary Size in Words",
       main = "")

# -----------------------------------------------
# create violin plot of CDI word total for each month
x16 <- CDIPQxs$wordtotal[CDIPQxs$camos == 16]
x17 <- CDIPQxs$wordtotal[CDIPQxs$camos == 17]
x18 <- CDIPQxs$wordtotal[CDIPQxs$camos == 18]
x19 <- CDIPQxs$wordtotal[CDIPQxs$camos == 19]
x20 <- CDIPQxs$wordtotal[CDIPQxs$camos == 20]
x21 <- CDIPQxs$wordtotal[CDIPQxs$camos == 21]
x22 <- CDIPQxs$wordtotal[CDIPQxs$camos == 22]
x23 <- CDIPQxs$wordtotal[CDIPQxs$camos == 23]
x24 <- CDIPQxs$wordtotal[CDIPQxs$camos == 24]
x25 <- CDIPQxs$wordtotal[CDIPQxs$camos == 25]
x26 <- CDIPQxs$wordtotal[CDIPQxs$camos == 26]
x27 <- CDIPQxs$wordtotal[CDIPQxs$camos == 27]
x28 <- CDIPQxs$wordtotal[CDIPQxs$camos == 28]
x29 <- CDIPQxs$wordtotal[CDIPQxs$camos == 29]
x30 <- CDIPQxs$wordtotal[CDIPQxs$camos == 30]

vioplot(x16, x17, x18, x19, x20, x21, x22, x23,
        x24, x25, x26, x27, x28, x29, x30, 
        names = c("16", "17", "18", "19", "20", "21", "22", 
                  "23", "24", "25", "26", "27", "28", "29", "30"), 
        col = "linen")  
# unsure how to add axis labels; xlab and ylab don't work

# -----------------------------------------------
# create density ridgeline plots of word totals by age
ggplot(CDIPQxs, aes(x = wordtotal, y = camos)) + 
  geom_density_ridges(scale = 0.85) +
  scale_x_continuous(limits = c(0, 700)) +
  theme_few() +
  scale_colour_few(palette = "dark") 

# -----------------------------------------------
# create notched boxplots of CDI grammatical complexity score for each month
boxplot(gctotal ~ camos, data = CDIPQxs, 
        notch=TRUE, col = "linen", range=1.5,
        # main = "Grammatical Complexity by Age", 
        xlab = "Age in Months", 
        ylab = "Grammatical Complexity Score")

# calculate mean complexity scores for each age group
mean_gc <- aggregate(gctotal ~ camos, CDIPQxs, mean)

# add mean complexity scores to graph
points(mean_gc$gctotal, pch = "*")

# -----------------------------------------------
# create box-percentile plots of CDI grammatical complexity for each month
bpplot(split(CDIPQxs$gctotal, CDIPQxs$camos),
       xlab = "Age in Months", 
       ylab = "Grammatical Complexity Score",
       main = "")

# -----------------------------------------------
# create violin plot of CDI grammatical complexity for each month
y16 <- CDIPQxs$gctotal[CDIPQxs$camos == 16]
y17 <- CDIPQxs$gctotal[CDIPQxs$camos == 17]
y18 <- CDIPQxs$gctotal[CDIPQxs$camos == 18]
y19 <- CDIPQxs$gctotal[CDIPQxs$camos == 19]
y20 <- CDIPQxs$gctotal[CDIPQxs$camos == 20]
y21 <- CDIPQxs$gctotal[CDIPQxs$camos == 21]
y22 <- CDIPQxs$gctotal[CDIPQxs$camos == 22]
y23 <- CDIPQxs$gctotal[CDIPQxs$camos == 23]
y24 <- CDIPQxs$gctotal[CDIPQxs$camos == 24]
y25 <- CDIPQxs$gctotal[CDIPQxs$camos == 25]
y26 <- CDIPQxs$gctotal[CDIPQxs$camos == 26]
y27 <- CDIPQxs$gctotal[CDIPQxs$camos == 27]
y28 <- CDIPQxs$gctotal[CDIPQxs$camos == 28]
y29 <- CDIPQxs$gctotal[CDIPQxs$camos == 29]
y30 <- CDIPQxs$gctotal[CDIPQxs$camos == 30]

vioplot(y16, y17, y18, y19, y20, y21, y22, y23,
        y24, y25, y26, y27, y28, y29, y30, 
        names = c("16", "17", "18", "19", "20", "21", "22", 
                  "23", "24", "25", "26", "27", "28", "29", "30"), 
        col = "linen")  
# unsure how to add axis labels; xlab and ylab don't work

# delete temporary objects ------------------
rm(region_tbl)
rm(statsnz)
rm(tbl_bosex)
rm(bosex_counts)
rm(bosex_props)
rm(bosex_rowcounts)
rm(NZbirths)
rm(region_props)
rm(region_totals)
rm(tbl_agesex1)
rm(tbl_agesex2)

sessionInfo()
